Projeler/Dolunay_Reels_Kapak/requirements.txt
notion-client==2.2.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-genai>=1.0.0
Pillow>=10.2.0
rembg==2.0.53
python-dotenv==1.0.1
requests>=2.31.0
